Refactor: highlightText 분리
